[Feat] 🍺 영업점 목록 api 정렬 구현

+++ src/main/java/com/hanaro/schedule_hanaro/customer/dto/response/AtmInfoDto.java
@@ -20,7 +20,7 @@ public record AtmInfoDto(
 	@JsonProperty("branch_type")
 	String branchType,
 	@JsonProperty("distance")
-	long distance
+	double distance
 ) {
 	public static AtmInfoDto of(
 		Long id,
@@ -30,7 +30,7 @@ public record AtmInfoDto(
 		String address,
 		String businessHours,
 		String branchType,
-		Long distance
+		double distance
 	) {
 		return AtmInfoDto.builder()
 			.id(id)
